Move ProcessSpec interface to include dir
