+++ src/platform/backends/qemu/dnsmasq_process_spec.h
@@ -21,7 +21,7 @@
 #include <multipass/ip_address.h>
 #include <multipass/optional.h>
 #include <multipass/path.h>
-#include <shared/linux/process_spec.h>
+#include <multipass/process_spec.h>
 
 #include <string>
