+++ src/platform/backends/shared/linux/linux_process.cpp
@@ -16,7 +16,6 @@
  */
 
 #include "linux_process.h"
-#include "process_spec.h"
 
 #include <multipass/logging/log.h>
